small refactor
moved part of the code to a utils module. and cleanup some parts of the code
