+++ custom_components/polygonal_zones/__init__.py
@@ -1,12 +1,8 @@
 """The Polygonal Zones integration."""
 
-import json
 import logging
 
-import aiohttp
-import pandas as pd
-from shapely.geometry import shape
-
+from .utils import get_zones
 from .const import (CONF_REGISTERED_ENTITIES, CONF_ZONES_URL, DATA_ZONES,
                     DATA_ZONES_URL, DOMAIN, PLATFORM)
 
@@ -31,6 +27,7 @@ async def async_setup_entry(hass, entry):
     await hass.config_entries.async_forward_entry_setups(entry, [PLATFORM])
     return True
 
+
 async def async_unload_entry(hass, entry):
     """Unload a config entry."""
     unload_ok = True
@@ -43,22 +40,3 @@ async def async_unload_entry(hass, entry):
         del hass.data[DOMAIN][entry.entry_id]
 
     return unload_ok
-
-
-async def load_data(uri: str) -> str:
-    async with aiohttp.ClientSession() as session:
-        async with session.get(uri) as response:
-            return await response.text()
-
-
-async def get_zones(uri: str) -> pd.DataFrame:
-    data = await load_data(uri)
-    data = json.loads(data)
-
-    zones = []
-    for i, feature in enumerate(data["features"]):
-        geometry = shape(feature["geometry"])
-        properties = feature["properties"]
-        zones.append({"geometry": geometry, **properties})
-
-    return pd.DataFrame(zones)
